+++ src/main/java/io/secuconnect/client/api/LoyaltyCardsApi.java
@@ -10,9 +10,11 @@ import io.secuconnect.client.ProgressRequestBody;
 import io.secuconnect.client.ProgressResponseBody;
 import com.google.gson.reflect.TypeToken;
 import java.io.IOException;
+import io.secuconnect.client.model.CardPin;
 import io.secuconnect.client.model.LoyaltyCardsList;
 import io.secuconnect.client.model.LoyaltyCardsProductModel;
 import io.secuconnect.client.model.ProductExceptionPayload;
+import io.secuconnect.client.model.ResultBoolean;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -110,12 +112,12 @@ public class LoyaltyCardsApi {
      * Assign loyalty card to specific user account
      * @param loyaltyCardId Loyalty card id (required)
      * @param generalAccountId General account id (required)
-     * @param loyaltyCardPin Loyalty card pin for the specific card
-     * @return Object
+     * @param body Loyalty card pin for the specific card
+     * @return ResultBoolean
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public Object assignUser(String loyaltyCardId, String generalAccountId, Object loyaltyCardPin) throws ApiException {
-        ApiResponse<Object> resp = assignUserWithHttpInfo(loyaltyCardId, generalAccountId, loyaltyCardPin);
+    public ResultBoolean assignUser(String loyaltyCardId, String generalAccountId, CardPin body) throws ApiException {
+        ApiResponse<ResultBoolean> resp = assignUserWithHttpInfo(loyaltyCardId, generalAccountId, body);
         return resp.getData();
     }
 
@@ -124,13 +126,13 @@ public class LoyaltyCardsApi {
      * Assign loyalty card to specific user account
      * @param loyaltyCardId Loyalty card id (required)
      * @param generalAccountId General account id (required)
-     * @param loyaltyCardPin Loyalty card pin for the specific card
-     * @return ApiResponse&lt;Object&gt;
+     * @param body Loyalty card pin for the specific card
+     * @return ApiResponse&lt;ResultBoolean&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<Object> assignUserWithHttpInfo(String loyaltyCardId, String generalAccountId, Object loyaltyCardPin) throws ApiException {
-        com.squareup.okhttp.Call call = assignUserValidateBeforeCall(loyaltyCardId, generalAccountId, loyaltyCardPin, null, null);
-        Type localVarReturnType = new TypeToken<Object>(){}.getType();
+    public ApiResponse<ResultBoolean> assignUserWithHttpInfo(String loyaltyCardId, String generalAccountId, CardPin body) throws ApiException {
+        com.squareup.okhttp.Call call = assignUserValidateBeforeCall(loyaltyCardId, generalAccountId, body, null, null);
+        Type localVarReturnType = new TypeToken<ResultBoolean>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
 
@@ -139,12 +141,12 @@ public class LoyaltyCardsApi {
      * Assign loyalty card to specific user account
      * @param loyaltyCardId Loyalty card id (required)
      * @param generalAccountId General account id (required)
-     * @param loyaltyCardPin Loyalty card pin for the specific card
+     * @param body Loyalty card pin for the specific card
      * @param callback The callback to be executed when the API call finishes
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public com.squareup.okhttp.Call assignUserAsync(String loyaltyCardId, String generalAccountId, Object loyaltyCardPin, final ApiCallback<Object> callback) throws ApiException {
+    public com.squareup.okhttp.Call assignUserAsync(String loyaltyCardId, String generalAccountId, CardPin body, final ApiCallback<ResultBoolean> callback) throws ApiException {
 
         ProgressResponseBody.ProgressListener progressListener = null;
         ProgressRequestBody.ProgressRequestListener progressRequestListener = null;
@@ -165,8 +167,8 @@ public class LoyaltyCardsApi {
             };
         }
 
-        com.squareup.okhttp.Call call = assignUserValidateBeforeCall(loyaltyCardId, generalAccountId, loyaltyCardPin, progressListener, progressRequestListener);
-        Type localVarReturnType = new TypeToken<Object>(){}.getType();
+        com.squareup.okhttp.Call call = assignUserValidateBeforeCall(loyaltyCardId, generalAccountId, body, progressListener, progressRequestListener);
+        Type localVarReturnType = new TypeToken<ResultBoolean>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;
     }
@@ -175,8 +177,8 @@ public class LoyaltyCardsApi {
      * @param count The number of items to return.
      * @param offset The position within the whole result set to start returning items (First element is at 0).
      * @param fields List of fields to include in the result. Nested properties can be accessed with this notation: prop1.prop2  Example: prop3,prop1.prop2
-     * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form property:condition.  *                  A condition may contain:  *                      - wildcard \&quot;*\&quot; for any number of characters  *                      - wildcard \&quot;?\&quot; for one character  *                      - ranges in the form [value TO value]  *  *                  Single expressions may combined by &#39;AND&#39;, &#39;OR&#39;, &#39;NOT&#39; operators and parenthesis &#39;(&#39;, &#39;)&#39; for grouping.  *                  Property names can be nested like \&quot;prop1.prop2\&quot;.  *                  Example: (NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])  *                  
-     * @param sort String with comma separated pairs of field:order (e.g. contact.surname:asc,contact.comapnyname:desc). Result set will be sorted by included fields, in ascending &#39;asc&#39;, or descending &#39;dsc&#39; order.
+     * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form property:condition.  *                   A condition may contain:  *                       - wildcard \&quot;*\&quot; for any number of characters  *                       - wildcard \&quot;?\&quot; for one character  *                       - ranges in the form [value TO value]  *  *                   Single expressions may combined by &#x27;AND&#x27;, &#x27;OR&#x27;, &#x27;NOT&#x27; operators and parenthesis &#x27;(&#x27;, &#x27;)&#x27; for grouping.  *                   Property names can be nested like \&quot;prop1.prop2\&quot;.  *                   Example: (NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])  *                   
+     * @param sort String with comma separated pairs of field:order (e.g. contact.surname:asc,contact.comapnyname:desc). Result set will be sorted by included fields, in ascending &#x27;asc&#x27;, or descending &#x27;dsc&#x27; order.
      * @param progressListener Progress listener
      * @param progressRequestListener Progress request listener
      * @return Call to execute
@@ -500,11 +502,11 @@ public class LoyaltyCardsApi {
      * Remove assigned loyalty card from specific user account
      * @param loyaltyCardId Loyalty card id (required)
      * @param generalAccountId General account id (required)
-     * @return Object
+     * @return ResultBoolean
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public Object removeAssignedUser(String loyaltyCardId, String generalAccountId) throws ApiException {
-        ApiResponse<Object> resp = removeAssignedUserWithHttpInfo(loyaltyCardId, generalAccountId);
+    public ResultBoolean removeAssignedUser(String loyaltyCardId, String generalAccountId) throws ApiException {
+        ApiResponse<ResultBoolean> resp = removeAssignedUserWithHttpInfo(loyaltyCardId, generalAccountId);
         return resp.getData();
     }
 
@@ -513,12 +515,12 @@ public class LoyaltyCardsApi {
      * Remove assigned loyalty card from specific user account
      * @param loyaltyCardId Loyalty card id (required)
      * @param generalAccountId General account id (required)
-     * @return ApiResponse&lt;Object&gt;
+     * @return ApiResponse&lt;ResultBoolean&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<Object> removeAssignedUserWithHttpInfo(String loyaltyCardId, String generalAccountId) throws ApiException {
+    public ApiResponse<ResultBoolean> removeAssignedUserWithHttpInfo(String loyaltyCardId, String generalAccountId) throws ApiException {
         com.squareup.okhttp.Call call = removeAssignedUserValidateBeforeCall(loyaltyCardId, generalAccountId, null, null);
-        Type localVarReturnType = new TypeToken<Object>(){}.getType();
+        Type localVarReturnType = new TypeToken<ResultBoolean>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
 
@@ -531,7 +533,7 @@ public class LoyaltyCardsApi {
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public com.squareup.okhttp.Call removeAssignedUserAsync(String loyaltyCardId, String generalAccountId, final ApiCallback<Object> callback) throws ApiException {
+    public com.squareup.okhttp.Call removeAssignedUserAsync(String loyaltyCardId, String generalAccountId, final ApiCallback<ResultBoolean> callback) throws ApiException {
 
         ProgressResponseBody.ProgressListener progressListener = null;
         ProgressRequestBody.ProgressRequestListener progressRequestListener = null;
@@ -553,7 +555,7 @@ public class LoyaltyCardsApi {
         }
 
         com.squareup.okhttp.Call call = removeAssignedUserValidateBeforeCall(loyaltyCardId, generalAccountId, progressListener, progressRequestListener);
-        Type localVarReturnType = new TypeToken<Object>(){}.getType();
+        Type localVarReturnType = new TypeToken<ResultBoolean>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;
     }